requirements.txt
Pillow>=9.0.0
opencv-python>=4.8.0
numpy>=1.21.0
Flask>=2.3.0
Werkzeug>=2.3.0
Flask-CORS>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.25.0
